Re-organized more code
