+++ src/decko/__init__.py
@@ -3,7 +3,6 @@ from .app import Decko
 
 # Function decorators
 from .decorators import deckorator
-from .decorators import stopwatch
 from .decorators import execute_if
 from .decorators import slower_than
 from .decorators import freeze
