+++ tests/decorators/test_decorators.py
@@ -8,7 +8,7 @@ import typing as t
 
 import src.decko.decorators as fd
 
-from src.decko.debug import try_except
+from src.decko.debug import try_except, stopwatch
 from tests.common.classes import Props
 from src.decko.immutable import ImmutableError
 
@@ -245,27 +245,6 @@ def test_class_method_decoration():
         return a + b
 
     assert add(1, 2) == 3, "Something is wrong ..."
-
-
-@pytest.mark.parametrize("iter_count", [
-    4, 8, 12
-])
-def test_stopwatch(iter_count):
-    time_elapsed_arr = []
-
-    def callback(val):
-        time_elapsed_arr.append(val)
-        return val
-
-    @fd.stopwatch(callback)
-    def create_list(n):
-        return list(range(n))
-
-    for i in range(iter_count):
-        create_list(1000000 * i)
-
-    assert len(time_elapsed_arr) == iter_count, \
-        f"Should have {len(time_elapsed_arr)} items"
 
 
 @pytest.mark.parametrize("input_data",
@@ -520,7 +499,7 @@ def test_multiple_decoration():
     the same results
     """
 
-    @fd.stopwatch(print)
+    @stopwatch(print)
     @try_except((ValueError, ), print)
     def add(a, b):
         return a + b
